Create README.md
Added a README.md file to the project. It contains instructions on how to use this project as well as the answer to the final question that was a part of this Code Challenge. May have to return to this file if the formatting I attempted doesn't work as I expected.